packages/core/pyproject.toml
[project]
name = "muspi-merol"
dynamic = ["version"]
description = "CLI utilities for personal use"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "rich~=13.9.4",
    "typer-slim~=0.12.5",
]

[project.scripts]
m = "m.cli.main:app"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.build]
includes = ["m"]
excludes = ["m/.store"]

[tool.pdm.version]
source = "file"
path = "m/version.py"
